Migrate contents of Docs -> DocsPageTemplate
